Reorganize API key logic; remove redundant tests.
