+++ middleware/security.py
@@ -2,34 +2,9 @@ from http import HTTPStatus
 from flask_jwt_extended import get_jwt_identity, verify_jwt_in_request
 from flask_restx import abort
 
-from database_client.database_client import DatabaseClient
 from database_client.helper_functions import get_db_client
-from middleware.access_logic import (
-    get_api_key_from_request_header,
-)
-from middleware.exceptions import (
-    InvalidAPIKeyException,
-    InvalidAuthorizationHeaderException,
-)
 from middleware.enums import PermissionsEnum
 from middleware.primary_resource_logic.permissions_logic import PermissionsManager
-
-INVALID_API_KEY_MESSAGE = "Please provide an API key in the request header in the 'Authorization' key with the format 'Basic <api_key>'"
-
-
-def check_api_key_associated_with_user(db_client: DatabaseClient, api_key: str) -> None:
-    user_identifiers = db_client.get_user_by_api_key(api_key)
-    if user_identifiers is None:
-        abort(HTTPStatus.UNAUTHORIZED, "Invalid API Key")
-
-
-def check_api_key() -> None:
-    try:
-        api_key = get_api_key_from_request_header()
-        db_client = get_db_client()
-        check_api_key_associated_with_user(db_client, api_key)
-    except (InvalidAPIKeyException, InvalidAuthorizationHeaderException):
-        abort(code=HTTPStatus.UNAUTHORIZED, message=INVALID_API_KEY_MESSAGE)
 
 
 def check_permissions(permission: PermissionsEnum) -> None:
