+++ resources/CreateTestUserWithElevatedPermissions.py
@@ -18,7 +18,7 @@ from flask import request
 from flask_restx import fields, abort
 
 from middleware.enums import PermissionsEnum
-from middleware.primary_resource_logic.login_queries import create_api_key_for_user
+from middleware.primary_resource_logic.api_key_logic import create_api_key_for_user
 from middleware.primary_resource_logic.user_queries import (
     user_post_results,
     UserRequestDTO,
