+++ tests/integration/test_api_key.py
@@ -1,15 +1,11 @@
 """Integration tests for /api_key endpoint"""
 
-from http import HTTPStatus
-
-from database_client.database_client import DatabaseClient
 from resources.ApiKey import API_KEY_ROUTE
 from tests.conftest import dev_db_client, flask_client_with_db
 from tests.helper_scripts.helper_functions import (
     create_test_user_api,
 )
 from tests.helper_scripts.run_and_validate_request import run_and_validate_request
-from tests.helper_scripts.simple_result_validators import check_response_status
 
 
 def test_api_key_post(flask_client_with_db, dev_db_client):
@@ -31,3 +27,5 @@ def test_api_key_post(flask_client_with_db, dev_db_client):
     assert new_user_info.api_key == response_json.get(
         "api_key"
     ), "API key returned not aligned with user API key in database"
+
+
